feat(statistics): add all cluster statistics
