+++ docs/swagger/swagger.json
@@ -13434,6 +13434,55 @@
                 }
             }
         },
+        "/v1/proxy/cluster/{cluster}/custom/statistics.system/v1/all": {
+            "get": {
+                "security": [
+                    {
+                        "JWT": []
+                    }
+                ],
+                "description": "获取集群级别统计",
+                "consumes": [
+                    "application/json"
+                ],
+                "produces": [
+                    "application/json"
+                ],
+                "tags": [
+                    "Agent.V1"
+                ],
+                "summary": "获取集群级别统计",
+                "parameters": [
+                    {
+                        "type": "string",
+                        "description": "cluster",
+                        "name": "cluster",
+                        "in": "path",
+                        "required": true
+                    }
+                ],
+                "responses": {
+                    "200": {
+                        "description": "counter",
+                        "schema": {
+                            "allOf": [
+                                {
+                                    "$ref": "#/definitions/handlers.ResponseStruct"
+                                },
+                                {
+                                    "type": "object",
+                                    "properties": {
+                                        "Data": {
+                                            "$ref": "#/definitions/statistics.ClusterStatistics"
+                                        }
+                                    }
+                                }
+                            ]
+                        }
+                    }
+                }
+            }
+        },
         "/v1/proxy/cluster/{cluster}/custom/statistics.system/v1/resources": {
             "get": {
                 "security": [
@@ -30794,6 +30843,20 @@
                 "used": {
                     "description": "集群资源的真实使用量",
                     "$ref": "#/definitions/v1.ResourceList"
+                }
+            }
+        },
+        "statistics.ClusterStatistics": {
+            "type": "object",
+            "properties": {
+                "resources": {
+                    "$ref": "#/definitions/statistics.ClusterResourceStatistics"
+                },
+                "version": {
+                    "type": "string"
+                },
+                "workloads": {
+                    "$ref": "#/definitions/statistics.ClusterWorkloadStatistics"
                 }
             }
         },
